rename events to scheduler
